Move `verifyCapabilityChain` to base purpose class and clean docs.
