+++ lib/CapabilityDelegation.js
@@ -301,6 +301,15 @@ module.exports = class CapabilityDelegation extends CapabilityProofPurpose {
     return {capability: {...document, proof}};
   }
 
+  async _runChecksBeforeChainVerification() {
+    /* Note: Here we create a signal to be sent to `_verifyCapabilityChain`
+    that the capability delegation proof for the tail has already been
+    verified (to avoid it being reverified). We will compute the full
+    `verifyResult` in `_runChecksAfterChainVerification` once we have verified
+    the parent capability. */
+    return {capabilityChainMeta: [{verifyResult: {}}]};
+  }
+
   async _runChecksAfterChainVerification({
     capabilityChainMeta, dereferencedChain, proof, validateOptions
   }) {
@@ -330,11 +339,15 @@ module.exports = class CapabilityDelegation extends CapabilityProofPurpose {
     const {verifiedParentCapability} = this;
     if(verifiedParentCapability) {
       // simple case, just validate against parent and return, we have been
-      // called from within a chain verification
+      // called from within a chain verification and can short circuit proof
+      // validation
       return this._validateAgainstParent({
         proof, verifiedParentCapability, validateOptions
       });
     }
+
+    // no short-circuit possible, we've just started validating the proof
+    // from root => tail
   }
 
   async _validateAgainstParent({
