+++ lib/CapabilityInvocation.js
@@ -411,6 +411,11 @@ module.exports = class CapabilityInvocation extends CapabilityProofPurpose {
         `Expected target (${expectedTarget}) does not match ` +
         `invocation target (${invocationTarget}).`);
     }
+
+    // return no capability delegation verify results yet; the tail's
+    // capability delegation proof must be verified via
+    // `_verifyCapabilityChain`
+    return {capabilityChainMeta: []};
   }
 
   async _runChecksAfterChainVerification({
