+++ lib/utils.js
@@ -3,7 +3,6 @@
  */
 'use strict';
 
-const jsigs = require('jsonld-signatures');
 const {MAX_CHAIN_LENGTH, ZCAP_CONTEXT_URL} = require('./constants');
 
 // FIXME: use `exports.foo` directly for improved destructuring w/ ESM import
@@ -423,327 +422,6 @@ api.dereferenceCapabilityChain = async ({
   return {dereferencedChain};
 };
 
-/**
- * A capability chain inspection function.
- * @typedef {function} InspectCapabilityChain
- * @param {CapabilityChainDetails}
- * @returns {InspectResult}
- */
-
-/**
- * A capability to inspect. The capability is compacted into the security
- * context. Only the required fields are shown here, a capability will contain
- * additional properties.
- * @typedef {object} Capability
- * @property {string} id - The ID of the capability.
- * @property {string} controller - The controller of the capability.
- */
-
-/**
- * The capability to inspect.
- * @typedef {object} CapabilityChainDetails
- * @property {Capability[]} capabilityChain - The capabilities in the chain.
- * @property {InspectChainResult[]} capabilityChainMeta - The results returned
- *   from jsonld-signatures verify for each capability in the chain. Each
- *   object contains `{verifyResult}` where each `verifyResult` is an
- *   `InspectChainResult`.
- */
-
-/**
- * The result of a capability chain inspection.
- * @typedef {object} InspectChainResult
- * @property {VerifyResult} verifyResult - The capability verify result.
- */
-
-/**
- * The result of running jsonld-signature's verify method.
- * @typedef {object} VerifyResult
- * @property {boolean} verified - `true` if all the checked proofs were
- *   successfully verified.
- * @property {VerifyProofResult[]} results - The verify results for each
- *   delegation proof.
- */
-
-/**
- * The result of verifying a capability delegation proof.
- * @typedef {object} VerifyProofResult
- * @property {VerifyProofPurposeResult} proofPurposeResult - The result from
- *   verifying the capability delegation proof purpose.
- */
-
-/**
- * The result of verifying a capability delegation proof purpose.
- * @typedef {object} VerifyProofPurposeResult
- * @property {string} delegator - The party that created the capability
- *   delegation proof, i.e., the party that delegated the capability.
- */
-
-/**
- * Verifies the given dereferenced capability chain. This involves ensuring
- * that the root zcap in the chain is as expected (for the endpoint where an
- * invocation or a simple chain chain is occurring) and that every other zcap
- * in the chain (including any invoked one), has been properly delegated.
- *
- * @param {class} CapabilityDelegation - The CapabilityDelegation class; this
- *   must be passed to avoid circular references in this module.
- * @param {boolean} [allowTargetAttenuation=false] - Allow the
- *   `invocationTarget` of a delegation chain to be increasingly restrictive
- *   based on a hierarchical RESTful URL structure.
- * @param {InspectChainResult[]} [capabilityChainMeta] - An optional array of
- *   results for inspecting the capability chain; this is passed when the
- *   capability delegation proof for the last capability in the dereferenced
- *   chain has already been verified, signaling that it should be not
- *   reverified within this function; all other verification results will be
- *   added to this array for passing to an optional capability inspection
- *   function.
- * @param {Date} [currentDate = new Date()] - The date used for comparison.
- *   when determining if a capability has expired.
- * @param {array} dereferencedChain - The dereferenced capability chain for
- *   `capability`, starting at the root capability and ending at `capability`.
- * @param {function} documentLoader - A configured jsonld documentLoader.
- * @param {object} expansionMap - A configured jsonld expansionMap.
- * @param {string|array} expectedRootCapability - The expected root
- *   capability for the delegation chain (this can be a single root
- *   capability ID expressed as a string or, if there is more than one
- *   acceptable root capability, several root capability IDs in an array.
- * @param {boolean} [requireChainDateMonotonicity=false] - Require the
- *   created dates on delegation proofs to be monotonically increasing
- *   forward in time.
- * @param {number} [maxDelegationTtl=Infinity] - The maximum time to live
- *   for a delegated zcap (as measured by the time difference between
- *   `expires`, which MUST be present if `maxDelegationTtl !== Infinity`, and
- *   `created` on the delegation proof.
- * @param {object|array} [suite] - The jsonld-signature suite(s) to use to
- *   verify the capability chain.
- *
- * @return {object} {verified, error, verifiedParentCapability}.
- */
-api.verifyCapabilityChain = async ({
-  CapabilityDelegation,
-  allowTargetAttenuation = false,
-  capabilityChainMeta = [],
-  currentDate = new Date(),
-  dereferencedChain,
-  documentLoader,
-  expansionMap,
-  expectedRootCapability,
-  maxDelegationTtl = Infinity,
-  requireChainDateMonotonicity = false,
-  suite
-}) => {
-  /* Note: We start verifying a capability from its root of trust (the
-  beginning or head of the capability chain) as this approach limits an
-  attacker's ability to waste our time and effort traversing from the tail
-  to the head. To prevent recursively repeating checks, we pass a
-  `verifiedParentCapability` each time we start verifying another capability
-  delegation proof in the capability chain.
-
-  Verification process is:
-
-  1. Verify the root capability ID matches an expected one.
-  2. For each capability `zcap` in the chain, verify the capability delegation
-    proof on `zcap` (if `capabilityChainMeta` has no precomputed result) and
-    that all of the delegation rules have been followed. */
-
-  /* Note: `verifiedParentCapability` will prevent repetitive checking of
-  the same segments of the chain (once a parent is verified, its chain is
-  not checked again when checking its children). */
-  let verifiedParentCapability = null;
-
-  try {
-    // 1. Verify the root capability ID matches an expected one.
-    const [root] = dereferencedChain;
-    if(!((Array.isArray(expectedRootCapability) &&
-      expectedRootCapability.includes(root.id)) ||
-      (typeof expectedRootCapability === 'string' &&
-        root.id === expectedRootCapability))) {
-      throw new Error(
-        `Expected root capability (${expectedRootCapability}) does not ` +
-        `match actual root capability (${root.id}).`);
-    }
-
-    // if the chain only has the root, exit early
-    if(dereferencedChain.length === 1) {
-      return {
-        verified: true,
-        verifiedParentCapability,
-        capabilityChainMeta
-      };
-    }
-
-    // 2. For each capability `zcap` in the chain, verify the capability
-    //   delegation proof on `zcap` and that the delegation rules have been
-    //   followed.
-    let parentAllowedAction;
-    let parentDelegationDate;
-    let parentExpirationDate;
-    let {invocationTarget: parentInvocationTarget} = root;
-
-    // track whether `capabilityChainMeta` needs its first result shifted to
-    // the end (if a result was present, it is for the last zcap, so move it
-    // to the end when we're done checking zcaps below)
-    const mustShift = capabilityChainMeta.length > 0;
-
-    // get all delegated capabilities (no root zcap since it has no delegation
-    // proof to check)
-    const delegatedCapabilities = dereferencedChain.slice(1);
-    for(let i = 0; i < delegatedCapabilities.length; ++i) {
-      const zcap = delegatedCapabilities[i];
-      verifiedParentCapability = delegatedCapabilities[i - 1] || root;
-
-      // verify proof on zcap if no result has been computed yet (one
-      // verify result will be present in `capabilityChainMeta` per
-      // delegated capability)
-      if(capabilityChainMeta.length < delegatedCapabilities.length) {
-        const verifyResult = await jsigs.verify(zcap, {
-          suite,
-          purpose: new CapabilityDelegation({
-            allowTargetAttenuation,
-            expectedRootCapability,
-            verifiedParentCapability,
-            maxDelegationTtl,
-            requireChainDateMonotonicity
-          }),
-          documentLoader,
-          expansionMap
-        });
-        if(!verifyResult.verified) {
-          throw verifyResult.error;
-        }
-        // delegation proof verified; save meta data for later inspection
-        capabilityChainMeta.push({verifyResult});
-      }
-
-      // ensure `allowedAction` is valid (compared against parent)
-      const {allowedAction} = zcap;
-      if(!_hasValidAllowedAction({allowedAction, parentAllowedAction})) {
-        throw new Error(
-          'The "allowedAction" in a delegated capability ' +
-          'must be equivalent or more restrictive than its parent.');
-      }
-
-      // ensure `invocationTarget` delegation is acceptable
-      const invocationTarget = api.getTarget({capability: zcap});
-      if(!api.isValidTarget({
-        invocationTarget,
-        baseInvocationTarget: parentInvocationTarget,
-        allowTargetAttenuation
-      })) {
-        if(allowTargetAttenuation) {
-          throw new Error(
-            `The "invocationTarget" in a delegated capability must be ` +
-            'equivalent or more restrictive than its parent.');
-        } else {
-          throw new Error(
-            'The "invocationTarget" in a delegated capability ' +
-            'must be equivalent to its parent.');
-        }
-      }
-
-      // verify expiration dates
-      // if the parent does not specify an expiration date, then any more
-      // restrictive expiration date is acceptable
-      let currentCapabilityExpirationDate;
-      if(zcap.expires !== undefined) {
-        // expires date has been previously validated, so just parse it
-        currentCapabilityExpirationDate = Date.parse(zcap.expires);
-      }
-
-      if(parentExpirationDate !== undefined ||
-        currentCapabilityExpirationDate !== undefined) {
-        // handle case where `expires` is set in the parent, but the child
-        // does not have `exires` or when the child has an expiration date
-        // greater than the parent.
-        if(currentCapabilityExpirationDate === undefined ||
-          currentCapabilityExpirationDate > parentExpirationDate) {
-          throw new Error('The `expires` property in a delegated ' +
-            'capability must be equivalent or more restrictive than its ' +
-            'parent.');
-        }
-        if(currentDate.getTime() > currentCapabilityExpirationDate) {
-          throw new Error('A capability in the delegation chain has expired.');
-        }
-      }
-
-      // get delegated date if necessary
-      let currentCapabilityDelegationDate;
-      const getDelegatedDate = requireChainDateMonotonicity ||
-        maxDelegationTtl < Infinity;
-      if(getDelegatedDate) {
-        const proofs = api.getDelegationProofs({capability: zcap});
-        // get earliest date from delegation proofs
-        for(const p of proofs) {
-          const created = Date.parse(p.created);
-          if(currentCapabilityDelegationDate === undefined) {
-            currentCapabilityDelegationDate = created;
-          } else if(currentCapabilityDelegationDate < created) {
-            currentCapabilityDelegationDate = created;
-          }
-        }
-      }
-
-      // FIXME: remove optionality here and always run this check
-      if(requireChainDateMonotonicity) {
-        // verify parent capability was not delegated after child
-        if(parentDelegationDate !== undefined &&
-          parentDelegationDate > currentCapabilityDelegationDate) {
-          throw new Error(
-            'A capability in the delegation chain was delegated before ' +
-            'its parent.');
-        }
-      }
-
-      // FIXME: disallow infinity as an option and set a reasonable default
-      if(maxDelegationTtl < Infinity) {
-        /* Note: Here we ensure zcap has a time-to-live (TTL) that is
-        sufficiently short. This is to prevent the use of zcaps that, when
-        revoked, will have to be stored for long periods of time. We have to
-        ensure:
-
-        1. The zcap has an expiration date.
-        2. The zcap's delegation date is not in the future (this also ensures
-          that the zcap's expiration date is not before its delegation date as
-          it would have triggered an expiration error in a previous check).
-        3. The zcap's current TTL is <= `maxDelegationTtl`
-        4. The zcap's TTL was never > `maxDelegationTtl`. */
-        if(zcap.expires === undefined) {
-          throw new Error(
-            'A delegated capability in the delegation chain does not have ' +
-            'an expiration date.');
-        }
-        if(currentCapabilityDelegationDate > currentDate) {
-          throw new Error(
-            'A delegated capability in the delegation chain was delegated ' +
-            'in the future.');
-        }
-        const currentTtl = currentCapabilityExpirationDate - currentDate;
-        const maxTtl = currentCapabilityExpirationDate -
-          currentCapabilityDelegationDate;
-        if(currentTtl > maxDelegationTtl || maxTtl > maxDelegationTtl) {
-          throw new Error(
-            'A delegated capability in the delegation chain has a time to ' +
-            'live that is too long.');
-        }
-      }
-
-      parentAllowedAction = allowedAction;
-      parentExpirationDate = currentCapabilityExpirationDate;
-      parentDelegationDate = currentCapabilityDelegationDate;
-      parentInvocationTarget = invocationTarget;
-    }
-
-    // shift zcap verify result for last zcap to the end of meta array if
-    // necessary
-    if(mustShift) {
-      capabilityChainMeta.push(capabilityChainMeta.shift());
-    }
-
-    return {verified: true, verifiedParentCapability, capabilityChainMeta};
-  } catch(error) {
-    return {verified: false, error};
-  }
-};
-
 api.checkProofContext = ({proof}) => {
   // zcap context can appear anywhere in the array as it *is* protected
   const {'@context': ctx} = proof;
@@ -752,6 +430,25 @@ api.checkProofContext = ({proof}) => {
     throw new Error(
       `Missing required capability proof context ("${ZCAP_CONTEXT_URL}").`);
   }
+};
+
+api.hasValidAllowedAction = ({allowedAction, parentAllowedAction}) => {
+  // if the parent's `allowedAction` is `undefined`, then any more restrictive
+  // action is allowed in the child
+  if(!parentAllowedAction) {
+    return true;
+  }
+
+  if(Array.isArray(parentAllowedAction)) {
+    // parent's `allowedAction` must include every one from child's
+    if(Array.isArray(allowedAction)) {
+      return allowedAction.every(a => parentAllowedAction.includes(a));
+    }
+    return parentAllowedAction.includes(allowedAction);
+  }
+
+  // require exact match
+  return (parentAllowedAction === allowedAction);
 };
 
 function _validateCapability({capability, expectRoot}) {
@@ -822,21 +519,59 @@ function _validateCapability({capability, expectRoot}) {
   }
 }
 
-function _hasValidAllowedAction({allowedAction, parentAllowedAction}) {
-  // if the parent's `allowedAction` is `undefined`, then any more restrictive
-  // action is allowed in the child
-  if(!parentAllowedAction) {
-    return true;
-  }
+// documentation typedefs
 
-  if(Array.isArray(parentAllowedAction)) {
-    // parent's `allowedAction` must include every one from child's
-    if(Array.isArray(allowedAction)) {
-      return allowedAction.every(a => parentAllowedAction.includes(a));
-    }
-    return parentAllowedAction.includes(allowedAction);
-  }
+/**
+ * A capability chain inspection function.
+ * @typedef {function} InspectCapabilityChain
+ * @param {CapabilityChainDetails}
+ * @returns {InspectResult}
+ */
 
-  // require exact match
-  return (parentAllowedAction === allowedAction);
-}
+/**
+ * A capability to inspect. The capability is compacted into the security
+ * context. Only the required fields are shown here, a capability will contain
+ * additional properties.
+ * @typedef {object} Capability
+ * @property {string} id - The ID of the capability.
+ * @property {string} controller - The controller of the capability.
+ */
+
+/**
+ * The capability to inspect.
+ * @typedef {object} CapabilityChainDetails
+ * @property {Capability[]} capabilityChain - The capabilities in the chain.
+ * @property {InspectChainResult[]} capabilityChainMeta - The results returned
+ *   from jsonld-signatures verify for each capability in the chain. Each
+ *   object contains `{verifyResult}` where each `verifyResult` is an
+ *   `InspectChainResult`.
+ */
+
+/**
+ * The result of a capability chain inspection.
+ * @typedef {object} InspectChainResult
+ * @property {VerifyResult} verifyResult - The capability verify result.
+ */
+
+/**
+ * The result of running jsonld-signature's verify method.
+ * @typedef {object} VerifyResult
+ * @property {boolean} verified - `true` if all the checked proofs were
+ *   successfully verified.
+ * @property {VerifyProofResult[]} results - The verify results for each
+ *   delegation proof.
+ */
+
+/**
+ * The result of verifying a capability delegation proof.
+ * @typedef {object} VerifyProofResult
+ * @property {VerifyProofPurposeResult} proofPurposeResult - The result from
+ *   verifying the capability delegation proof purpose.
+ */
+
+/**
+ * The result of verifying a capability delegation proof purpose.
+ * @typedef {object} VerifyProofPurposeResult
+ * @property {string} delegator - The party that created the capability
+ *   delegation proof, i.e., the party that delegated the capability.
+ */
